migrate additionalDisplay into LCEM
